requirements.txt
# GX量化 Web 统一平台 v2.0 - 依赖包

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# 数据库
sqlalchemy>=2.0.0
sqlite3 (built-in)

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# 量化分析
scikit-learn>=1.3.0
scipy>=1.11.0

# AI/大模型
requests>=2.31.0
aiohttp>=3.9.0
tenacity>=8.2.0

# 任务调度
apscheduler>=3.10.0

# 日志
loguru>=0.7.0
python-dotenv>=1.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# 开发工具
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# 性能分析
py-spy>=0.3.14

# 工具库
tqdm>=4.66.0
colorlog>=6.7.0
